+++ module-web/build.gradle
@@ -1,0 +1,16 @@
+bootJar {
+    enabled=true
+}
+
+jar {
+    enabled=true
+}
+
+dependencies {
+    implementation project(':module-api')
+    implementation project(':module-common')
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.flywaydb:flyway-core'
+    runtimeOnly 'com.h2database:h2'
+    testImplementation 'io.rest-assured:rest-assured:3.3.0'
+}
