+++ module-web/src/main/java/kitchenpos/product/ProductRestController.java
@@ -1,0 +1,40 @@
+package kitchenpos.product;
+
+import kitchenpos.product.application.ProductService;
+import kitchenpos.product.dto.ProductRequest;
+import kitchenpos.product.dto.ProductResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.net.URI;
+import java.util.List;
+import java.util.stream.Collectors;
+
+@RestController
+public class ProductRestController {
+    private final ProductService productService;
+
+    public ProductRestController(final ProductService productService) {
+        this.productService = productService;
+    }
+
+    @PostMapping("/api/products")
+    public ResponseEntity<ProductResponse> create(@RequestBody final ProductRequest productRequest) {
+        final ProductResponse created = productService.create(productRequest);
+        final URI uri = URI.create("/api/products/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @GetMapping("/api/products")
+    public ResponseEntity<List<ProductResponse>> list() {
+        List<ProductResponse> productResponses = productService.list();
+        return ResponseEntity.ok()
+                .body(productResponses)
+                ;
+    }
+}
