+++ module-web/src/main/java/kitchenpos/table/TableRestController.java
@@ -1,0 +1,60 @@
+package kitchenpos.table;
+
+import kitchenpos.table.application.TableService;
+import kitchenpos.table.dto.OrderTableRequest;
+import kitchenpos.table.dto.OrderTableResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.net.URI;
+import java.util.List;
+import java.util.stream.Collectors;
+
+@RestController
+public class TableRestController {
+    private final TableService tableService;
+
+    public TableRestController(final TableService tableService) {
+        this.tableService = tableService;
+    }
+
+    @PostMapping("/api/tables")
+    public ResponseEntity<OrderTableResponse> create(@RequestBody final OrderTableRequest orderTableRequest) {
+        final OrderTableResponse created = tableService.create(orderTableRequest);
+        final URI uri = URI.create("/api/tables/" + created.getId());
+        return ResponseEntity.created(uri)
+                .body(created)
+                ;
+    }
+
+    @GetMapping("/api/tables")
+    public ResponseEntity<List<OrderTableResponse>> list() {
+
+        List<OrderTableResponse> orderTableResponses = tableService.list();
+        return ResponseEntity.ok()
+                .body(orderTableResponses)
+                ;
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/empty")
+    public ResponseEntity<OrderTableResponse> changeEmpty(
+            @PathVariable final Long orderTableId,
+            @RequestBody final OrderTableRequest orderTableRequest
+    ) {
+        OrderTableResponse orderTableResponse = tableService.changeEmpty(orderTableId, orderTableRequest);
+        return ResponseEntity.ok()
+                .body(orderTableResponse)
+                ;
+    }
+
+    @PutMapping("/api/tables/{orderTableId}/number-of-guests")
+    public ResponseEntity<OrderTableResponse> changeNumberOfGuests(
+            @PathVariable final Long orderTableId,
+            @RequestBody final OrderTableRequest orderTableRequest
+    ) {
+        OrderTableResponse orderTableResponse = tableService.changeNumberOfGuests(orderTableId, orderTableRequest);
+        return ResponseEntity.ok()
+                .body(orderTableResponse)
+                ;
+    }
+}
