+++ module-web/src/main/resources/db/migration/V2__Insert_default_data.sql
