+++ module-web/src/test/java/kitchenpos/order/OrderFactory.java
@@ -1,0 +1,41 @@
+package kitchenpos.order;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import kitchenpos.order.dto.OrderRequest;
+import org.springframework.http.MediaType;
+
+
+public class OrderFactory {
+
+    public static ExtractableResponse<Response> 주문_생성_요청(OrderRequest params) {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().post("/api/orders")
+                .then().log().all().
+                extract();
+    }
+
+    public static ExtractableResponse<Response> 주문_조회_요청() {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().get("/api/orders")
+                .then().log().all().
+                extract();
+    }
+
+    public static ExtractableResponse<Response> 주문_상태변경_요청(OrderRequest params, Long orderId) {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().put("/api/orders/"+orderId+"/order-status")
+                .then().log().all().
+                extract();
+    }
+
+}
