+++ module-web/src/test/java/kitchenpos/tablegroup/TableGroupFactory.java
@@ -1,0 +1,38 @@
+package kitchenpos.tablegroup;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import kitchenpos.tablegroup.dto.TableGroupRequest;
+import kitchenpos.tablegroup.dto.TableGroupResponse;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class TableGroupFactory {
+
+    public static ExtractableResponse<Response> 주문그룹테이블_생성_요청(TableGroupRequest params) {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().post("/api/table-groups")
+                .then().log().all().
+                extract();
+    }
+
+    public static ExtractableResponse<Response> 주문그룹테이블_삭제_요청(Long tableGroupId) {
+        return RestAssured
+                .given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().delete("/api/table-groups/"+tableGroupId)
+                .then().log().all().
+                extract();
+    }
+
+    public static TableGroupResponse 주문그룹테이블이_생성됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+        return response.as(TableGroupResponse.class);
+    }
+}
