--- src/main/java/kitchenpos/ui/MenuRestController.java
@@ -1,44 +1,0 @@
-package kitchenpos.ui;
-
-import kitchenpos.menu.application.MenuService;
-import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.dto.MenuRequest;
-import kitchenpos.menu.dto.MenuResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@RestController
-public class MenuRestController {
-    private final MenuService menuService;
-
-    public MenuRestController(final MenuService menuService) {
-        this.menuService = menuService;
-    }
-
-    @PostMapping("/api/menus")
-    public ResponseEntity<MenuResponse> create(@RequestBody final MenuRequest menuRequest) {
-        final Menu created = menuService.create(menuRequest);
-        final URI uri = URI.create("/api/menus/" + created.getId());
-        return ResponseEntity.created(uri)
-                .body(MenuResponse.from(created))
-                ;
-    }
-
-    @GetMapping("/api/menus")
-    public ResponseEntity<List<MenuResponse>> list() {
-        List<Menu> menus = menuService.list();
-        List<MenuResponse> menuResponses = menus.stream()
-                                                .map(MenuResponse::from)
-                                                .collect(Collectors.toList());
-        return ResponseEntity.ok()
-                .body(menuResponses)
-                ;
-    }
-}
