--- src/main/java/kitchenpos/ui/OrderRestController.java
@@ -1,51 +1,0 @@
-package kitchenpos.ui;
-
-import kitchenpos.order.application.OrderService;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.dto.OrderRequest;
-import kitchenpos.order.dto.OrderResponse;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
-import java.util.List;
-import java.util.stream.Collectors;
-
-@RestController
-public class OrderRestController {
-    private final OrderService orderService;
-
-    public OrderRestController(final OrderService orderService) {
-        this.orderService = orderService;
-    }
-
-    @PostMapping("/api/orders")
-    public ResponseEntity<OrderResponse> create(@RequestBody final OrderRequest orderRequest) {
-        final Order created = orderService.create(orderRequest);
-        final URI uri = URI.create("/api/orders/" + created.getId());
-        return ResponseEntity.created(uri)
-                .body(OrderResponse.from(created))
-                ;
-    }
-
-    @GetMapping("/api/orders")
-    public ResponseEntity<List<OrderResponse>> list() {
-
-        List<Order> orders = orderService.list();
-        List<OrderResponse> orderResponses = orders.stream()
-                                                    .map(OrderResponse::from)
-                                                    .collect(Collectors.toList());
-        return ResponseEntity.ok()
-                .body(orderResponses)
-                ;
-    }
-
-    @PutMapping("/api/orders/{orderId}/order-status")
-    public ResponseEntity<OrderResponse> changeOrderStatus(
-            @PathVariable final Long orderId,
-            @RequestBody final OrderRequest orderRequest
-    ) {
-        OrderResponse orderResponse = OrderResponse.from(orderService.changeOrderStatus(orderId, orderRequest));
-        return ResponseEntity.ok(orderResponse);
-    }
-}
