--- src/test/java/kitchenpos/application/acceptance/product/ProductFactory.java
@@ -1,29 +1,0 @@
-package kitchenpos.application.acceptance.product;
-
-import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-import kitchenpos.product.dto.ProductRequest;
-import org.springframework.http.MediaType;
-
-public class ProductFactory {
-
-    public static ExtractableResponse<Response> 상품_생성_요청(ProductRequest params) {
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(params)
-                .when().post("/api/products")
-                .then().log().all().
-                extract();
-    }
-
-    public static ExtractableResponse<Response> 상품_조회_요청() {
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/api/products")
-                .then().log().all().
-                extract();
-    }
-}
